refactor: MainActivity를 MovieListActivity로 변경
