+++ app/src/main/java/woowacourse/movie/movie/MovieListActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.main
+package woowacourse.movie.movie
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
@@ -8,17 +8,17 @@ import androidx.core.view.WindowInsetsCompat
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
-import woowacourse.movie.main.adapter.MoviesAdapter
+import woowacourse.movie.movie.adapter.MoviesAdapter
 import woowacourse.movie.reserve.ReserveActivity
 
-class MainActivity : AppCompatActivity(), MainContract.View {
+class MovieListActivity : AppCompatActivity(), MovieListContract.View {
     private val moviesView: RecyclerView by lazy { findViewById(R.id.rv_movies) }
-    private val presenter: MainPresenter = MainPresenter(this)
+    private val presenter: MovieListPresenter = MovieListPresenter(this)
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        setContentView(R.layout.activity_main)
+        setContentView(R.layout.activity_movie_list)
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
