+++ app/src/main/java/woowacourse/movie/movie/adapter/ViewType.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.main.adapter
+package woowacourse.movie.movie.adapter
 
 enum class ViewType {
     TYPE_MOVIE,
